Program options refactor for PGGigECam
- Works, but needs testing.
- Options need descriptions.
- Also, would be good to have mono vs color option and consequent
  specializations of color conversion calls.
- Added grasshopper manual to docs.

+++ src/frameserver/WebCam.h
@@ -49,7 +49,7 @@ public:
 private:
 
     // The webcam object
-    int64_t index_ {0};
+    size_t index_ {0};
     std::unique_ptr<cv::VideoCapture> cv_camera_;
 
     // frame generation clock
